Allow only DeliveryRemovedEvent in deleteDelivery method
